+++ tests/testthat/test-get_mcmc_samples.R
@@ -10,6 +10,21 @@ test_that("Default Bayesian hierarchical model can be build", {
                            compute_ess = TRUE, max_iter = 50000)
   expect_equal(class(test$samples), "mcmc")
 
+  test <- get_mcmc_samples(data = get_mock_data(),
+                           id_application = "application",
+                           id_voter = "voter",
+                           grade_variable = "num_grade",
+                           id_section = "panel",
+                           inits_type = "overdispersed",
+                           variables_to_sample = "specified",
+                           names_variables_to_sample =
+                             c("tau_application", "tau_voter", "sigma",
+                               "rank_theta", "tau_section"),
+                           n_chains = 4, n_adapt = 10000, n_iter = 10000,
+                           n_burnin = 10000, dont_bind = TRUE,
+                           compute_ess = TRUE, max_iter = 10000)
+  expect_equal(class(test$samples), "mcmc")
+
   # test <- get_mcmc_samples(data = get_mock_data(),
   #                          id_application = "application",
   #                          id_voter = "voter", id_section = "panel",
@@ -24,7 +39,7 @@ test_that("Default Bayesian hierarchical model can be build", {
                            inits_type = "overdispersed",
                            heterogeneous_residuals = TRUE,
                            n_chains = 4, n_adapt = 10000, n_iter = 10000,
-                           n_burnin = 10000, max_iter = 50000)
+                           n_burnin = 10000, max_iter = 10000)
   expect_equal(class(test$samples), "mcmc")
 
   test <- get_mcmc_samples(data = dat, id_application = "application",
@@ -32,15 +47,15 @@ test_that("Default Bayesian hierarchical model can be build", {
                            grade_variable = "num_grade",
                            heterogeneous_residuals = TRUE,
                            n_chains = 4, n_adapt = 10000, n_iter = 10000,
-                           n_burnin = 10000, max_iter = 50000)
+                           n_burnin = 10000, max_iter = 10000)
   expect_equal(class(test$samples), "mcmc")
 
   test <- get_mcmc_samples(data = dat, id_application = "application",
                            id_voter = "voter",
                            grade_variable = "num_grade",
                            inits_type = "overdispersed",
-                           n_chains = 4, n_adapt = 50000, n_iter = 10000,
-                           n_burnin = 10000, max_iter = 50000)
+                           n_chains = 4, n_adapt = 10000, n_iter = 10000,
+                           n_burnin = 10000, max_iter = 10000)
   expect_equal(class(test$samples), "mcmc")
 
   test <- get_mcmc_samples(data = dat, id_application = "application",
@@ -49,7 +64,7 @@ test_that("Default Bayesian hierarchical model can be build", {
                            point_scale = 6,
                            ordinal_scale = TRUE,
                            n_chains = 4, n_adapt = 10000, n_iter = 10000,
-                           n_burnin = 10000, max_iter = 50000)
+                           n_burnin = 10000, max_iter = 10000)
   expect_equal(class(test$samples), "mcmc")
 
   test <- get_mcmc_samples(data = dat, id_application = "application",
@@ -59,7 +74,7 @@ test_that("Default Bayesian hierarchical model can be build", {
                            point_scale = 6,
                            ordinal_scale = TRUE,
                            n_chains = 4, n_adapt = 10000, n_iter = 10000,
-                           n_burnin = 10000, max_iter = 50000)
+                           n_burnin = 10000, max_iter = 10000)
   expect_equal(class(test$samples), "mcmc")
 
 
@@ -71,6 +86,6 @@ test_that("Default Bayesian hierarchical model can be build", {
                            point_scale = 6,
                            ordinal_scale = TRUE,
                            n_chains = 4, n_adapt = 10000, n_iter = 10000,
-                           n_burnin = 10000, max_iter = 50000)
+                           n_burnin = 10000, max_iter = 10000)
   expect_equal(class(test$samples), "mcmc")
 })